\name{politantheme-package}
\alias{politantheme-package}
\alias{politantheme}
\docType{package}
\title{
Politantheme
}
\description{
Provides a theme for the politan.ch blog as well as some functions.
}
\details{
\tabular{ll}{
Package: \tab politantheme\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-13\cr
License: \tab What license is it under?\cr
}
bez_br_index \cr gamma_mean \cr gamma_mode \cr politantheme_col \cr theme_politan
}
\author{
Th. W.
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
